Create Parser.java to parse user command
